+++ routes/api.php
@@ -7,11 +7,11 @@ use Gildsmith\Product\Controllers\Attribute\AttributeDeleteController;
 use Gildsmith\Product\Controllers\Attribute\AttributeFindController;
 use Gildsmith\Product\Controllers\Attribute\AttributeIndexController;
 use Gildsmith\Product\Controllers\Attribute\AttributeUpdateController;
-use Gildsmith\Product\Controllers\Attribute\Value\ValueCreateController;
-use Gildsmith\Product\Controllers\Attribute\Value\ValueDeleteController;
-use Gildsmith\Product\Controllers\Attribute\Value\ValueFindController;
-use Gildsmith\Product\Controllers\Attribute\Value\ValueIndexController;
-use Gildsmith\Product\Controllers\Attribute\Value\ValueUpdateController;
+use Gildsmith\Product\Controllers\AttributeValue\AttributeValueCreateController;
+use Gildsmith\Product\Controllers\AttributeValue\AttributeValueDeleteController;
+use Gildsmith\Product\Controllers\AttributeValue\AttributeValueFindController;
+use Gildsmith\Product\Controllers\AttributeValue\AttributeValueIndexController;
+use Gildsmith\Product\Controllers\AttributeValue\AttributeValueUpdateController;
 use Gildsmith\Product\Controllers\Product\ProductCreateController;
 use Gildsmith\Product\Controllers\Product\ProductDeleteController;
 use Gildsmith\Product\Controllers\Product\ProductFindController;
@@ -34,17 +34,17 @@ Route::prefix('products')->group(function () {
 Route::prefix('attributes')->group(function () {
     Route::get('/', AttributeIndexController::class);
     Route::post('/', AttributeCreateController::class);
-    Route::get('/{code}', AttributeFindController::class);
-    Route::put('/{code}', AttributeUpdateController::class);
-    Route::patch('/{code}', AttributeUpdateController::class);
-    Route::delete('/{code}', AttributeDeleteController::class);
+    Route::get('/{attribute}', AttributeFindController::class);
+    Route::put('/{attribute}', AttributeUpdateController::class);
+    Route::patch('/{attribute}', AttributeUpdateController::class);
+    Route::delete('/{attribute}', AttributeDeleteController::class);
 
     Route::prefix('{attribute}/values')->group(function () {
-        Route::get('/', ValueIndexController::class);
-        Route::post('/', ValueCreateController::class);
-        Route::get('/{value}', ValueFindController::class);
-        Route::put('/{value}', ValueUpdateController::class);
-        Route::patch('/{value}', ValueUpdateController::class);
-        Route::delete('/{value}', ValueDeleteController::class);
+        Route::get('/', AttributeValueIndexController::class);
+        Route::post('/', AttributeValueCreateController::class);
+        Route::get('/{value}', AttributeValueFindController::class);
+        Route::put('/{value}', AttributeValueUpdateController::class);
+        Route::patch('/{value}', AttributeValueUpdateController::class);
+        Route::delete('/{value}', AttributeValueDeleteController::class);
     });
 });
