+++ src/Controllers/AttributeValue/AttributeValueFindController.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Gildsmith\Product\Controllers\Attribute\Value;
+namespace Gildsmith\Product\Controllers\AttributeValue;
 
 use Gildsmith\Contract\Product\AttributeValueInterface;
 use Gildsmith\Support\Facades\Product;
 use Illuminate\Routing\Controller;
 
-class ValueFindController extends Controller
+class AttributeValueFindController extends Controller
 {
     public function __invoke(string $attribute, string $value): ?AttributeValueInterface
     {
