+++ src/Controllers/AttributeValue/AttributeValueIndexController.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Gildsmith\Product\Controllers\Attribute\Value;
+namespace Gildsmith\Product\Controllers\AttributeValue;
 
 use Gildsmith\Support\Facades\Product;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Collection;
 
-class ValueIndexController extends Controller
+class AttributeValueIndexController extends Controller
 {
     public function __invoke(string $attribute): Collection
     {
