+++ src/Controllers/AttributeValue/AttributeValueUpdateController.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Gildsmith\Product\Controllers\Attribute\Value;
+namespace Gildsmith\Product\Controllers\AttributeValue;
 
 use Gildsmith\Contract\Product\AttributeValueInterface;
 use Gildsmith\Support\Facades\Product;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 
-class ValueUpdateController extends Controller
+class AttributeValueUpdateController extends Controller
 {
     public function __invoke(Request $request, string $attribute, string $value): AttributeValueInterface
     {
